Extract embeddedSummaryListitem to its own file
This was causing a dependency cycle so its best to extract it
