+++ server/utils/checkYourAnswersUtils/index.ts
@@ -1,11 +1,12 @@
 import { TemporaryAccommodationApplication } from '@approved-premises/api'
 import { HtmlItem, SummaryListItem, Task, TextItem } from '@approved-premises/ui'
 
-import paths from '../paths/apply'
+import paths from '../../paths/apply'
 
-import { forPagesInTask } from './applicationUtils'
-import reviewSections from './reviewUtils'
-import { escape, formatLines } from './viewUtils'
+import reviewSections from '../reviewUtils'
+import { formatLines } from '../viewUtils'
+import { embeddedSummaryListItem } from './embeddedSummaryListItem'
+import { forPagesInTask } from '../applicationUtils'
 
 const checkYourAnswersSections = (application: TemporaryAccommodationApplication) =>
   reviewSections(application, getTaskResponsesAsSummaryListItems)
@@ -32,29 +33,6 @@ export const getTaskResponsesAsSummaryListItems = (
   return items
 }
 
-const embeddedSummaryListItem = (answers: Array<Record<string, unknown>>): string => {
-  let response = ''
-
-  answers.forEach(answer => {
-    response += '<dl class="govuk-summary-list govuk-summary-list--embedded">'
-    Object.keys(answer).forEach(key => {
-      response += `
-      <div class="govuk-summary-list__row govuk-summary-list__row--embedded">
-        <dt class="govuk-summary-list__key govuk-summary-list__key--embedded">
-         ${escape(key)}
-        </dt>
-        <dd class="govuk-summary-list__value govuk-summary-list__value--embedded">
-        ${escape(answer[key] as string)}
-        </dd>
-      </div>
-      `
-    })
-    response += '</dl>'
-  })
-
-  return response
-}
-
 const summaryListItemForResponse = (
   key: string,
   value: TextItem | HtmlItem,
@@ -79,4 +57,4 @@ const summaryListItemForResponse = (
   }
 }
 
-export { checkYourAnswersSections, embeddedSummaryListItem }
+export { checkYourAnswersSections }
